pyproject.toml
[tool.poetry]
name = "financer-api"
version = "0.1.0"
description = ""
authors = ["Felipe Quecole <felipequecole@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
websockets = "^12.0"
pydantic-settings = "^2.1.0"
nest-asyncio = "^1.6.0"
fastapi-pagination = "^0.12.14"


[tool.poetry.group.dev.dependencies]
pyway = "^0.3.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
